+++ lib/datarockets/style.rb
@@ -4,6 +4,7 @@ require "datarockets/style/formatter/todo_list_formatter"
 require "datarockets/style/version"
 
 require "datarockets/style/cop/layout/array_alignment_extended"
+require "datarockets/style/cop/style/nested_interpolation"
 
 module Datarockets
   # Datarickors sharable config